requirements.txt
Flask==0.10.1
psycopg2==2.7.1
Flask-SQLAlchemy==2.2
pandas==0.19.1
oauth2client==4.0.0
httplib2==0.9.2
urllib3==1.20
google-api-python-client==1.6.2
configparser==3.5.0
requests==2.13.0
